Add stdin-reading versions of day2
One uses "blocking reads" (BRK followed by `.Console/read DEI` —
`.Console/read` is not really blocking but the preceding BRK
will suspend the VM until a byte is available), the
other uses `.Console/vector` and could in theory do other things
while processing the input (but just goes to sleep instead).

Both versions share the issue that there doesn't seem to be a
way to detect the end of standard input. If you're blocked on
a BRK when stdin gets closed, the VM will just exit, regardless
of whether you're using `.Console/vector`.

For now, to signal the end of the input, there needs to be a
line starting with `.` after all real input lines, e.g.:

  $ (cat day2.txt; echo .) | uxncli ./day2-blockingread.rom